metric: add datadriven test for histogram buckets
Release note: None
